agregando validaciones a clase de servicio Reserva

+++ reservavuelos/src/main/java/com/proyecto/reservavuelos/service/ReservaService.java
@@ -21,6 +21,11 @@ public class ReservaService {
 
     public Reserva crearReserva(Reserva reserva) {
         // Aquí puedes realizar validaciones antes de guardar la reserva si es necesario
+
+        if (reserva == null) {
+            throw new IllegalArgumentException("La reserva no puede ser null");
+        }
+
         return reservaRepository.save(reserva);
     }
 
@@ -33,6 +38,12 @@ public class ReservaService {
     }
 
     public Reserva actualizarReserva(Long id, Reserva reservaActualizada) {
+        if (id == null) {
+            throw new IllegalArgumentException("El ID no puede ser null");
+        }
+        if (reservaActualizada == null) {
+            throw new IllegalArgumentException("La reserva actualizada no puede ser null");
+        }
         Reserva reservaExistente = reservaRepository.findById(id).orElse(null);
         if (reservaExistente != null) {
             reservaActualizada.setId(id); // Aseguramos que el ID no cambie
@@ -42,7 +53,11 @@ public class ReservaService {
         return null; // La reserva no existe
     }
 
+
     public boolean eliminarReserva(Long id) {
+        if (id == null) {
+            throw new IllegalArgumentException("El ID no puede ser null");
+        }
         Reserva reserva = reservaRepository.findById(id).orElse(null);
         if (reserva != null) {
             reservaRepository.delete(reserva);
